fix: skip process.exit in dev

+++ packages/kit/src/vite/index.js
@@ -75,6 +75,9 @@ function kit() {
 	/** @type {import('types').ManifestData} */
 	let manifest_data;
 
+	/** @type {boolean} */
+	let is_build;
+
 	/**
 	 * @type {{
 	 *   build_dir: string;
@@ -117,6 +120,7 @@ function kit() {
 		async config(config, { command }) {
 			vite_config = config;
 			svelte_config = await load_config();
+			is_build = command === 'build';
 
 			paths = {
 				build_dir: `${svelte_config.kit.outDir}/build`,
@@ -124,7 +128,7 @@ function kit() {
 				client_out_dir: `${svelte_config.kit.outDir}/output/client/${svelte_config.kit.appDir}`
 			};
 
-			if (command === 'build') {
+			if (is_build) {
 				process.env.VITE_SVELTEKIT_APP_VERSION = svelte_config.kit.version.name;
 				process.env.VITE_SVELTEKIT_APP_VERSION_FILE = `${svelte_config.kit.appDir}/version.json`;
 				process.env.VITE_SVELTEKIT_APP_VERSION_POLL_INTERVAL = `${svelte_config.kit.version.pollInterval}`;
@@ -184,11 +188,13 @@ function kit() {
 		},
 
 		buildStart() {
-			rimraf(paths.build_dir);
-			mkdirp(paths.build_dir);
+			if (is_build) {
+				rimraf(paths.build_dir);
+				mkdirp(paths.build_dir);
 
-			rimraf(paths.output_dir);
-			mkdirp(paths.output_dir);
+				rimraf(paths.output_dir);
+				mkdirp(paths.output_dir);
+			}
 		},
 
 		async writeBundle(_options, bundle) {
@@ -316,7 +322,7 @@ function kit() {
 		},
 
 		closeBundle() {
-			if (svelte_config.kit.prerender.enabled) {
+			if (is_build && svelte_config.kit.prerender.enabled) {
 				// this is necessary to close any open db connections, etc.
 				// TODO: prerender in a subprocess so we can exit in isolation
 				// https://github.com/sveltejs/kit/issues/5306
